Integrateur Runge-Kutta

+++ exerciceP10/general/Classe_Vecteur/Vecteur.cc
@@ -49,7 +49,7 @@ void Vecteur::correctif_0() {
  }
 
 void Vecteur::modulo2pi() {
-    for (int i(0); i<dim() ; ++i) {
+    for (size_t i(0); i<dim() ; ++i) {
         while (coeff(i)>=2.0*M_PI) {
             set_coord(i,coeff(i)-2.0*M_PI);
         }  // pour obtenir les angles modulo 2 pi
